+++ bin/load-benchmark-data
@@ -10,7 +10,7 @@ export K6_STATSD_ADDR=${K6_STATSD_ADDR:-"graphite-statsd:$K6_STATSD_PORT"}
 
 export CONJUR_ADMIN_API_KEY=$(./bin/cli conjur user rotate-api-key | tail -n 1 | tr -d '[:space:]')
 export APPLIANCE_MASTER_URL=${APPLIANCE_MASTER_URL:-"https://conjur-master.mycompany.local:443"}
-export APPLIANCE_FOLLOWER_URL=${APPLIANCE_FOLLOWER_URL:-"https://conjur-follower.mycompany.local:443"}
+export APPLIANCE_READ_URL=${APPLIANCE_READ_URL:-"https://conjur-follower.mycompany.local:443"}
 export CONJUR_ACCOUNT="demo"
 export CONJUR_IDENTITY="admin"
 export K6_FLAGS="--insecure-skip-tls-verify"
@@ -84,7 +84,7 @@ EOF
 #
 load_policy_set() {
   check_env "APPLIANCE_MASTER_URL"
-  check_env "APPLIANCE_FOLLOWER_URL"
+  check_env "APPLIANCE_READ_URL"
   check_env "CONJUR_ADMIN_API_KEY"
   check_env "POLICY_SET"
